pyproject.toml
[tool.poetry]
name = "project4_backend"
version = "0.1.0"
description = ""
authors = ["rodages <zanasmakarov@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.12"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
psycopg2 = "^2.9.3"
djangorestframework-simplejwt = "^5.1.0"
gunicorn = "^20.1.0"
whitenoise = "^6.0.0"
dj-database-url = "^0.5.0"
dj-rest-auth = "^2.2.4"
django-environ = "^0.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.950"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
